perf(webapp): componetization spendTool view
